Fix metrics implementation correctness issues
DATABASE:
- Change engagement_rate type from FLOAT to NUMERIC(6,5) for precision

UI:
- Replace !== null with != null for metrics visibility checks (safe for undefined)
- Includes wrapper condition and all individual metric checks
- Prevents calling toLocaleString() on undefined values

API:
- Add ISO timestamp validation for posted_at using Date.parse()
- Reject invalid timestamp formats
- Preserve existing validation logic for other fields

+++ docs/MARKETING_METRICS_MIGRATION.sql
@@ -8,7 +8,7 @@ ADD COLUMN likes INTEGER,
 ADD COLUMN comments INTEGER,
 ADD COLUMN shares INTEGER,
 ADD COLUMN saves INTEGER,
-ADD COLUMN engagement_rate FLOAT,
+ADD COLUMN engagement_rate NUMERIC(6,5),
 ADD COLUMN posted_at TIMESTAMPTZ,
 ADD COLUMN platform_post_id TEXT;
